web/secret_door/chal/challenge/requirements.txt
Flask==3.0.3
PyJWT==2.8.0
flask-mysqldb==2.0.0